tapback: Remove physical-device support
Moving to an NBD-only approach, physical-device is getting populated
with a major:minor of 0:0.  That is because we want to place the nbd
unix socket path, /run/blktap-control/nbd$pid.$minor, in
physical-device-path.  The block scripts use stat -L -c %t:%T "$1",
which only works on block or char device nodes.  Running it against the
unix socket returns 0:0.

physical-device-path has everything we want - the usable nbd server
path, pid and "minor".  Rely on that being set to setup the tapdisk
connection.

Dropping physical-device also avoids the issue that looking up by
"minor" would be incorrect since it would always be zero.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>