tapback: Add -l/--libxl compatibility mode
libxl and xapi/xenopsd have different ordering requirements, and the
libxl required changes are not readily compatible with xapi.

Add a flag to allow running in libxl compatibility mode to bodge around
the issue.  The flag has to be passed from the parent tapback process
into the per-domain ones to ensure it is respected.

Some nearby hard tabs are converted to spaces.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>